fix: make tests pass

+++ CA5/src/main/java/ir/ut/ie/contracts/PagedResponse.java
@@ -1,7 +1,7 @@
 package ir.ut.ie.contracts;
 
 public record PagedResponse<TItem> (
-    int totalItems,
+    long totalItems,
     int offset,
     int limit,
     TItem[] items
